+++ src/main/coerce.c
@@ -165,7 +165,7 @@ int attribute_hidden IntegerFromString(SEXP x, int *warn)
                 return INT_MIN;
             }
             else
-                return xdouble;
+                return (int)xdouble;
         }
         else
             *warn |= WARN_NA;
