+++ src/main/localecharset.c
@@ -695,7 +695,7 @@ char *locale2charset(const char *locale)
             {
                 if (charset[3] != '-')
                 {
-                    for (i = strlen(charset) - 3; 0 < i; i--)
+                    for (i = (int)strlen(charset) - 3; 0 < i; i--)
                         charset[i + 1] = charset[i];
                     charset[3] = '-';
                 }
